+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.2.1 (unreleased)
+2.3.0 (unreleased)
 ------------------
 
 - Add method ``tool.runImportStepFromText``.